+++ CHANGELOG.md
@@ -14,6 +14,7 @@ in the CfP!
 ## Unreleased
 
 - [#1524](https://github.com/babashka/babashka/issues/1524): Remove dynamic builds for linux-aarch64 ([@lispyclouds](https://github.com/lispyclouds))
+- [#1577](https://github.com/babashka/babashka/issues/1557): Add support for `babashka.process/exec` after namespace reload of `babashka.process` ([@lread](https://github.com/lread)) 
 
 ## 1.3.179 (2023-04-26)